+++ src/api/get_image.rs
@@ -3,27 +3,25 @@ use crate::api::ImageInfo;
 
 use super::send_api_request;
 
-use log::error;
+use anyhow::Error as anyhow_err;
 use reqwest::Method;
-use std::process::exit;
+use std::io::{Error, ErrorKind};
 
-pub async fn get_image(c: ImgurHandle, image: &str) -> Result<ImageInfo, String> {
+pub async fn get_image(c: ImgurHandle, image: &str) -> Result<ImageInfo, anyhow_err> {
     let uri = api_url!(format!("image/{image}"));
-    let res = send_api_request(&c, Method::GET, uri, None)
-        .await
-        .unwrap_or_else(|e| {
-            error!("send api request: {e}");
-            exit(1)
-        });
+    let res = send_api_request(&c, Method::GET, uri, None).await?;
 
     let status = res.status();
 
     if status.is_client_error() || status.is_server_error() {
-        Err(format!(
-            "server returned non-successful status code = {status}."
-        ))
+        let err = Error::new(
+            ErrorKind::Other,
+            format!("server returned non-successful status code = {status}"),
+        );
+
+        Err(anyhow_err::from(err))
     } else {
-        let content: ImageInfo = res.json().await.map_err(|e| e.to_string())?;
+        let content: ImageInfo = res.json().await.map_err(anyhow_err::new)?;
         Ok(content)
     }
 }