+++ src/api/upload_image.rs
@@ -5,34 +5,33 @@ use crate::api::{
 
 use super::send_api_request;
 
-use log::error;
+use anyhow::Error as anyhow_err;
 use reqwest::Method;
-use std::{collections::HashMap, process::exit};
+use std::{
+    collections::HashMap,
+    io::{Error, ErrorKind},
+};
 
-pub async fn upload_image(c: ImgurHandle, image: &str) -> Result<ImageInfo, String> {
+pub async fn upload_image(c: ImgurHandle, image: &str) -> Result<ImageInfo, anyhow_err> {
     let mut form = HashMap::new();
 
     form.insert("image", image.to_string());
 
     let uri = api_url!("image");
-    let res = send_api_request(&c, Method::POST, uri, Some(form))
-        .await
-        .unwrap_or_else(|e| {
-            error!("send api request: {e}");
-            exit(1)
-        });
+    let res = send_api_request(&c, Method::POST, uri, Some(form)).await?;
 
     let status = res.status();
 
     if status.is_client_error() || status.is_server_error() {
-        let body = res.text().await.map_err(|e| e.to_string())?;
+        let body = res.text().await.map_err(anyhow_err::new)?;
+        let err = Error::new(
+            ErrorKind::Other,
+            format!("server returned non-successful status code = {status}, body = {body}"),
+        );
 
-        Err(format!(
-            "server returned non-successful status code = {status}. body = {body}"
-        ))
+        Err(anyhow_err::from(err))
     } else {
-        let content: ImageInfo = res.json().await.map_err(|e| e.to_string())?;
-
+        let content: ImageInfo = res.json().await.map_err(anyhow_err::new)?;
         Ok(content)
     }
 }