+++ src/config/toml.rs
@@ -1,5 +1,6 @@
 use super::Config;
 
+use anyhow::Error;
 use dirs::config_dir;
 use log::{error, warn};
 use std::{
@@ -10,6 +11,8 @@ use std::{
 };
 use toml::from_str;
 
+const CONFIG_DIR: &str = "/imgurs/config.toml";
+
 pub fn parse() -> Config {
     toml().unwrap_or_else(|e| {
         warn!("Parse toml config: {e}! Creating config file...");
@@ -17,7 +20,7 @@ pub fn parse() -> Config {
         let default_config = include_str!(concat!("../../config.toml"));
 
         let sys_config_dir = config_dir().unwrap();
-        let config_dir = format!("{}/imgurs/config.toml", sys_config_dir.to_string_lossy());
+        let config_dir = format!("{}{CONFIG_DIR}", sys_config_dir.to_string_lossy());
         let config_path = Path::new(&config_dir);
 
         let prefix = config_path.parent().unwrap();
@@ -31,19 +34,19 @@ pub fn parse() -> Config {
 
         toml().unwrap_or_else(|e| {
             error!("parse toml config: {e}");
-            exit(2);
+            exit(3);
         })
     })
 }
 
-fn toml() -> Result<Config, String> {
+fn toml() -> Result<Config, Error> {
     let config_dir = config_dir().unwrap();
 
-    let file_dir: String = String::from(config_dir.to_string_lossy()) + "/imgurs/config.toml";
+    let file_dir: String = String::from(config_dir.to_string_lossy()) + CONFIG_DIR;
 
-    let toml_str = read_to_string(file_dir).map_err(|err| err.to_string())?;
+    let toml_str = read_to_string(file_dir).map_err(|e| e)?;
 
-    let decode = from_str(&toml_str).map_err(|err| err.to_string())?;
+    let decode = from_str(&toml_str).map_err(|e| e)?;
 
     Ok(decode)
 }